Add Twitter credential handling logic (#59)
This commit adds a logic to include Twitter API credentials in the configuration file. Lastly, I also
added some minor refactors to ensure that the everything is extensible. That is, rather than typing out each environment variable and prompts separately, I created an opts variable that contains an anonymous struct that is handled by lookupEnvVar() and promptString() methods.

Signed-off-by: Lester James V. Miranda <ljvmiranda@gmail.com>